backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Auth
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.7

# Recipe scraping (Mealie's core engine)
recipe-scrapers>=15.0.0
httpx>=0.27.0

# Dev
python-dotenv>=1.0.0
